generalize the scheduler beyond timers
